Fix layout shift in NavigationStack
The core problem, as identified by Abe, is that determining the final layout (including the app bar's height and visibility) requires preference values derived from composing the content itself. This creates a chicken-and-egg scenario.

We now render the main content in a `Box` with `currentContentAlpha` set to `0.0` initially.

And we're now tracking a state variable, `isTopBarMeasuredOnce', initially false, which is set to true in an `onGloballyPositionedInWindow` modifier on the top app bar. When `isTopBarMeasuredOnce` is true, we make the main content visible.